some code refactoring
