+++ src/js/components/App/App.js
@@ -1,59 +1,8 @@
-import React, { PureComponent } from 'react';
-import {
-  HashRouter as Router, Route, Switch, Redirect
-} from 'react-router-dom';
-import { isUserAuth } from '../../helpers';
-import Header from '../Header';
-import Footer from '../Footer';
-import Courses from '../Courses';
-import LoginPage from '../LoginPage';
-import ErrorPage from '../ErrorPage';
-import ProtectedRoute from '../ProtectedRoute';
+import React from 'react';
+import ContainerComponent from '../AppContainer';
 
-class App extends PureComponent {
-  state = {
-    isAuth: false
-  };
-
-  componentDidMount() {
-    this.setState({ isAuth: isUserAuth() });
-  }
-
-  render() {
-    const authProps = {
-      isAuth: this.state.isAuth,
-      setAuth: isLogin => this.setState({ isAuth: isLogin })
-    };
-    return (
-      <Router>
-        <Header
-          {...authProps}
-        />
-        <Switch>
-          <Route
-            path="/"
-            exact
-            {...authProps}
-          >
-            <Redirect to="/courses" {...authProps} />
-          </Route>
-          <Route
-            path="/login"
-            {...authProps}
-            component={props => <LoginPage {...authProps} {...props} />}
-          />
-          <Route
-            path="/courses"
-            component={props => <Courses {...authProps} {...props} />}
-            {...authProps}
-          />
-          <Route component={ErrorPage} />
-        </Switch>
-
-        <Footer />
-      </Router>
-    );
-  }
-}
+const App = () => (
+  <ContainerComponent />
+);
 
 export default App;
