+++ src/js/components/AppContainer/index.js
@@ -1,0 +1,3 @@
+import AppContainer from './AppContainer';
+
+export default AppContainer;
