--- src/js/components/CoursesList/template.js
@@ -1,27 +1,0 @@
-const template = [
-  {
-    title: 'Video course 1',
-    timing: '27 min',
-    date: '27.09.12',
-    text: 'vjbdsvjsbvfskjjdbvkjbvsvjvbdfjbvdhb vdhsb cds chdnb cvhdddsvsvxsv',
-    id: 1
-  },
-
-  {
-    title: 'Video course 2',
-    timing: '1 h 29 min',
-    date: '04.09.12',
-    text: 'nfjbnjdgfbnfnvkjfvjbfkjbkjsdeb nkjbnvjb nvjdfs',
-    id: 2
-  },
-
-  {
-    title: 'Video course 3',
-    timing: '4 h 244 min',
-    date: '13.09.12',
-    text: 'nfjbnjdgfbnfnvkjfvjbfkjbkjsdeb nkjbnvjb nvjdfs',
-    id: 3
-  }
-];
-
-export default template;
